requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
cssselect==1.1.0
flake8==3.9.2
idna==3.2
lxml==4.6.3
mccabe==0.6.1
parsel==1.6.0
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.12.0
python-decouple==3.4
requests==2.26.0
six==1.16.0
urllib3==1.26.6
w3lib==1.22.0
